Refactored parse_options, implementing argument checking

+++ src/master_process.c
@@ -7,22 +7,33 @@
 /* largely for preparation and clean-up.                              */
 /**********************************************************************/
 void
-master_process (int N_p, options_struct *opts) {
+master_process (int N_p, int argc, char *argv[]) {
   PTnode *root = NULL;
-  double res_nrm;
-  int lambda_index = opts->lambda_index;
+  options_struct opts;
   bool has_succeeded;
 
-  /* First verification: meaningful tree depth must be larger than 1 */
-  if (opts->max_depth<=1) {
-    printf ("master_process: Require max_depth > 1; max_depth = %d\n",
-            opts->max_depth);
+  /* Default options before passing */
+  initialize_options (&opts);
+  opts.verbose = 2;
+  opts.max_children = 0;
+  has_succeeded = parse_options (argc, argv, &opts);
+  /* Verification: meaningful tree depth must be larger than 1 */
+  if (!has_succeeded) {
+    printf ("master_process: Unable to parse parameter file\n");
     printf ("Terminating...\n");
     goto cleanup;
   }
 
-  /* Second verification: Allocate root_node successfully */
-  has_succeeded = create_root_node (&root, opts);
+  /* Verification: meaningful tree depth must be larger than 1 */
+  if (opts.max_depth<=1) {
+    printf ("master_process: Require max_depth > 1; max_depth = %d\n",
+            opts.max_depth);
+    printf ("Terminating...\n");
+    goto cleanup;
+  }
+
+  /* Verification: Allocate root_node successfully */
+  has_succeeded = create_root_node (&root, &opts);
   if (!has_succeeded) {
     printf ("master_process:");
     printf (" Failed to allocate memory for root node.\n");
@@ -30,8 +41,8 @@ master_process (int N_p, options_struct *opts) {
     goto cleanup;
   }
 
-  /* Third verification: Load initial point from disk */
-  has_succeeded = load_initial_coordinates (root, opts);
+  /* Verification: Load initial point from disk */
+  has_succeeded = load_initial_coordinates (root, &opts);
   if (!has_succeeded) {
     printf ("master_process:");
     printf (" Failed to read first point into root node.\n");
@@ -39,7 +50,9 @@ master_process (int N_p, options_struct *opts) {
     goto cleanup;
   }
 
-  if (opts->verbose>0) {
+  if (opts.verbose>0) {
+    double res_nrm;
+    int lambda_index = opts.lambda_index;
     printf ("master_process: Loaded first point from file.\n");
     compute_residual (root->N_dim, root->z, root->T_init);
     res_nrm = cblas_dnrm2 (root->N_dim-1, root->T_init, 1);
@@ -48,32 +61,39 @@ master_process (int N_p, options_struct *opts) {
     printf ("Initial lambda = %12.5g\n", root->z[lambda_index]);
   }
 
-  /* Fourth verification: Determine initial secant direction */
-  has_succeeded = initialize_secant (root, opts);
+  /* Verification: Determine initial secant direction */
+  has_succeeded = initialize_secant (root, &opts);
   if (!has_succeeded) {
     printf ("master_process:");
     printf (" Failed to determine initial secant direction.\n");
-    printf ("Maximum of %d iterations exceeded.\n", opts->max_iter);
-    printf ("Desired residual tolerance: %12.5e\n",opts->tol_residual);
+    printf ("Maximum of %d iterations exceeded.\n", opts.max_iter);
+    printf ("Desired residual tolerance: %12.5e\n",opts.tol_residual);
     printf ("Terminating.\n");
     goto cleanup;
   }
 
-  principal_pampac_loop (N_p, root, opts);
-  write_root_coordinates (root, opts);
+  principal_pampac_loop (N_p, root, &opts);
+  write_root_coordinates (root, &opts);
 
 cleanup:
-  if (opts->verbose>0)
+  if (opts.verbose>0)
     printf("master_process: Shutting down slave processes.\n");
   stop_slaves(N_p);
-  if (opts->verbose>0)
+  if (opts.verbose>0)
     printf("master_process: Cleaning up memory.\n");
-  /* z_init is a shallow copy in every node of the tree *except* the
-   * root node. As such, z_init must be *explicitly* deallocated
-   * prior to calling the function delete_tree. */
-  free (root->z_init);
-  root->z_init = NULL;
-  delete_tree (root);
-  root = NULL;
+  if (opts.input_filename!=NULL)
+    free (opts.input_filename);
+  if (opts.output_filename!=NULL)
+    free (opts.output_filename);
+  if (opts.scale_factors!=NULL)
+    free (opts.scale_factors);
+  if (root!=NULL) {
+    /* z_init is a shallow copy in every node of the tree *except* the
+    * root node. As such, z_init must be *explicitly* deallocated
+    * prior to calling the function delete_tree. */
+    if (root->z_init != NULL)
+      free (root->z_init);
+    delete_tree (root);
+  }
   return;
 }
